rename structure package name
