+++ src/main/java/com/baidu/hugegraph/api/graph/VertexAPI.java
@@ -25,11 +25,11 @@ import java.util.Map;
 
 import javax.ws.rs.core.MultivaluedHashMap;
 
-import com.baidu.hugegraph.api.graph.structure.BatchVertexRequest;
 import com.baidu.hugegraph.client.RestClient;
 import com.baidu.hugegraph.exception.NotAllCreatedException;
 import com.baidu.hugegraph.rest.RestResult;
 import com.baidu.hugegraph.structure.constant.HugeType;
+import com.baidu.hugegraph.structure.graph.BatchVertexRequest;
 import com.baidu.hugegraph.structure.graph.Vertex;
 import com.baidu.hugegraph.structure.graph.Vertices;
 import com.google.common.collect.ImmutableMap;
