+++ src/main/java/com/baidu/hugegraph/api/traverser/SingleSourceShortestPathAPI.java
@@ -23,10 +23,10 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 
 import com.baidu.hugegraph.api.graph.GraphAPI;
-import com.baidu.hugegraph.api.traverser.structure.WeightedPaths;
 import com.baidu.hugegraph.client.RestClient;
 import com.baidu.hugegraph.rest.RestResult;
 import com.baidu.hugegraph.structure.constant.Direction;
+import com.baidu.hugegraph.structure.traverser.WeightedPaths;
 import com.baidu.hugegraph.util.E;
 
 public class SingleSourceShortestPathAPI extends TraversersAPI {
