+++ src/main/java/com/baidu/hugegraph/driver/TraverserManager.java
@@ -42,16 +42,6 @@ import com.baidu.hugegraph.api.traverser.ShortestPathAPI;
 import com.baidu.hugegraph.api.traverser.SingleSourceShortestPathAPI;
 import com.baidu.hugegraph.api.traverser.VerticesAPI;
 import com.baidu.hugegraph.api.traverser.WeightedShortestPathAPI;
-import com.baidu.hugegraph.api.traverser.structure.CountRequest;
-import com.baidu.hugegraph.api.traverser.structure.CrosspointsRequest;
-import com.baidu.hugegraph.api.traverser.structure.CustomizedCrosspoints;
-import com.baidu.hugegraph.api.traverser.structure.CustomizedPaths;
-import com.baidu.hugegraph.api.traverser.structure.FusiformSimilarity;
-import com.baidu.hugegraph.api.traverser.structure.FusiformSimilarityRequest;
-import com.baidu.hugegraph.api.traverser.structure.PathsRequest;
-import com.baidu.hugegraph.api.traverser.structure.Ranks;
-import com.baidu.hugegraph.api.traverser.structure.WeightedPath;
-import com.baidu.hugegraph.api.traverser.structure.WeightedPaths;
 import com.baidu.hugegraph.client.RestClient;
 import com.baidu.hugegraph.structure.constant.Direction;
 import com.baidu.hugegraph.structure.graph.Edge;
@@ -61,6 +51,16 @@ import com.baidu.hugegraph.structure.graph.Path;
 import com.baidu.hugegraph.structure.graph.Shard;
 import com.baidu.hugegraph.structure.graph.Vertex;
 import com.baidu.hugegraph.structure.graph.Vertices;
+import com.baidu.hugegraph.structure.traverser.CountRequest;
+import com.baidu.hugegraph.structure.traverser.CrosspointsRequest;
+import com.baidu.hugegraph.structure.traverser.CustomizedCrosspoints;
+import com.baidu.hugegraph.structure.traverser.CustomizedPaths;
+import com.baidu.hugegraph.structure.traverser.FusiformSimilarity;
+import com.baidu.hugegraph.structure.traverser.FusiformSimilarityRequest;
+import com.baidu.hugegraph.structure.traverser.PathsRequest;
+import com.baidu.hugegraph.structure.traverser.Ranks;
+import com.baidu.hugegraph.structure.traverser.WeightedPath;
+import com.baidu.hugegraph.structure.traverser.WeightedPaths;
 import com.baidu.hugegraph.util.E;
 
 import static com.baidu.hugegraph.structure.constant.Traverser.DEFAULT_CAPACITY;
