+++ src/test/java/com/baidu/hugegraph/api/FusiformSimilarityApiTest.java
@@ -26,13 +26,13 @@ import java.util.stream.Collectors;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import com.baidu.hugegraph.api.traverser.structure.FusiformSimilarity;
-import com.baidu.hugegraph.api.traverser.structure.FusiformSimilarity.Similar;
-import com.baidu.hugegraph.api.traverser.structure.FusiformSimilarityRequest;
 import com.baidu.hugegraph.exception.ServerException;
 import com.baidu.hugegraph.structure.constant.Direction;
 import com.baidu.hugegraph.structure.constant.T;
 import com.baidu.hugegraph.structure.graph.Vertex;
+import com.baidu.hugegraph.structure.traverser.FusiformSimilarity;
+import com.baidu.hugegraph.structure.traverser.FusiformSimilarityRequest;
+import com.baidu.hugegraph.structure.traverser.FusiformSimilarity.Similar;
 import com.baidu.hugegraph.testutil.Assert;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
