+++ src/test/java/com/baidu/hugegraph/api/WeightedShortestPathApiTest.java
@@ -24,10 +24,10 @@ import java.util.List;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import com.baidu.hugegraph.api.traverser.structure.WeightedPath;
 import com.baidu.hugegraph.structure.constant.Direction;
 import com.baidu.hugegraph.structure.constant.T;
 import com.baidu.hugegraph.structure.graph.Vertex;
+import com.baidu.hugegraph.structure.traverser.WeightedPath;
 import com.baidu.hugegraph.testutil.Assert;
 import com.google.common.collect.ImmutableList;
